Move IconCache into qtUtils
